fixes #304 perf tool exits too quickly on throughput client
This adds a synchronization at exit step, where the server sends
a completion message to the client.  This permits the client to
to be sure that the server has got all the data.  There is a 5
second timeout, which should allow this to work (albeit with a
weird 5 second completion delay) with legacy servers.