fixes #444 Stop looking for asciidoctor by default
While here, we have actually cleaned this up and enabled documentation
building to work, for folks that want it.  We default to off, because
frankly folks should just use the website and be happy.

If docs are enabled, then both man pages and html are generated.
The man pages are noticably inferior (only so much you can do with
80 columns of text), but some people have indicated they prefer to
be able to type "man" something.