fixes #484 crashes in websocket transport fixes #490 posix_epdesc use-after-free bug fixes #489 Sanitizer based testing would help fixes #492 Numerous memory leaks found with sanitizer
This introduces support for compiler-based sanitizers when using
clang or gcc (and not on Windows).  See NNG_SANITIZER for possible
settings such as "thread" or "address".

Furthermore, we have fixed the issues we found with both the
thread and address sanitizers.  We believe that the thread issues
pointed to a low frequency use-after-free responsible for rare
crashes in some of the tests.

The tests generally have their timeouts doubled when running under
a sanitizer, to account for the extra long times that the sanitizer
can cause these to take.

While here, we also changed the compat_ws test to avoid a particularly
painful and time consuming DNS lookup, and we made the nngcat_unlimited
test a bit more robust by waiting before sending traffic.