fixes #496 nngcat tests fail in Xcode build
This avoids copying the tests, but runs them with bash from
the source directory, and passes the path to the built nngcat
binary to the script.

It also permits these to run if we find bash, so that systems
without bash won't run it and fail.

We still only do this on POSIX though, because Windows bash
is too broken to use for some of the tests.