fixes #508 nng_aio_free blocks during callback
This recycles the job structures, and so the demo seems
to work.  If you have sufficient concurrency, these aio
structures will ultimately leak (with the level of concurrency),
but for this demo that's fine.  (To keep them from leaking, you'd
walk the freelist and discard them just before exiting.  The
actual growth of the list should be quite small.)  There are
other strategies that could be used to avoid uncontrolled growth,
but they aren't useful for our demo purposes.  Also, when we have
a fix for 511, we can go ahead and just do a deferred discard.

As aios are used *heavily* in the code, the strategy of a deferred
cache may help reduce presssure on the heap, andd might be something
we want to explore for the core.