fixes #533 nni_aio_begin should not dispatch task on NNG_ECLOSED.
This changes nni_aio_begin so that it immediately terminates when
it encounters aio->a_closed, much like it does for aio->a_stop.
The semantic for nni_aio_close() is supposed to be like nni_aio_stop(),
but without blocking.

I suspect that this might be responsible for use-after-free bugs that
seem to have been rearing their head lately.