fixes #523 dialers could support multiple outstanding dial requests fixes #179 DNS resolution should be done at connect time
This makes multiple dialers work for TCP on POSIX.  The Windows
dialer code needs further work to support more than a single dialer.

This is not yet 100% complete, as it doesn't move all of the TCP code to the
new API yet.

POSIX TCP refactored.

This uses a rewhacked TCP that does not use the epdesc and pipedesc,
but instead has its own API that is specific TCP.  It moves DNS lookups
to the connect time, leading to improved fault tolerance and self-healing,
as it can react to changes in the resolution of DNS names.  (This may aid
in self-healing and loadbalancing as well, where the resolver returns
records in different orders.)

First swag and Windows TCP rewhack.