fixes #611 Memory Leaks under Windows fixes #622 incorrect assumptions about malloc(0)
Windows actually allocates an object of size zero when calling
malloc on size zero.  This is unusual behavior, and we just
add logic to work more like malloc on POSIX systems.

Other systems can return non-NULL objects to fixed pages here.
We think the best option here is to uniformly return NULL from
our APIs in these circumstances, and to include testing to validate
that.