fixes #620 Stress tests are too stressful in CI/CD
This converts the tests to use async I/O callbacks instead of
threads for running tests.  This should greatly reduce the
amount of pressure we apply to the system.  On macOS the
time to start up with a pressure of 500 is significantly less
than under the old system.  Plus, as we are no longer at the mercy
of the scheduler, we're far more likely to get a successful test.