fixes #648 REQ protocol can hang on close
Actually the problem was in socket core, in particular in the
shutdown code.  The socket shutdown is supposed to ensure that
no pipes were present on the socket, so that protocols need not
concern themselves with this.  The code unfortunately was busted,
due to an ordering problem compounded by a race condition.  This
fixes that, and changes the REQ protocol to avoid the blocking
condition altogether, and sprinkles a few assertions to validate
these rules are being adhered to.