fixes #653 Weird connect failures in dialer (multistress)
The POLLHUP (or rather EPOLLHUP) flag does not quite mean the same
thing in Linux, and we've seen random failures where we will sometimes
get this event on a socket that is freshly connected.  This might be
a bug in Linux, but it is easy enough to workaround -- we just don't
watch for it at all.