Make docs for old versions and tip point to the correct page
All the man page docs that are not the latest released version give a notice, saying that the current docs being viewed are not the latest, and offer a link to the latest.  This fixes the broken link; formerly it was pointing to nng.html, now it points to index.html.