fixes #720 fix for #715 still not right
The fix for #715 to address const is still not quite right.

In ISO C, the argv array is just char **. getopt() in POSIX uses
char *const[]. That part is right, but we should then not pass
const char ** in calls. Furthermore, really the optarg should also
not be constified.

This aligns us closer to getopt() and leads to the fewest problems.
This does represent a slight breaking change, but as old code should
still compile and run, we don't think we should bump the API version
for this change.  Furthermore, we don't think anyone else is actually
using this API yet.