Fix for possible overruns in fragsz.
While the fragment size itself fits within 16-bits, when we do math
on it (such as multiplying by a fragment count), we may need to calculate
values that exceed 16-bits (e.g. 1MB).  Therefore we should treat this
as a size_t, but cast it down to 16-bits only when we write it out to
the packet field.