fixes #682 Support for Chunked Transfer Coding
This is the client side only, although the work is structured to
support server applications.  The chunked API is for now private,
although the intent to is to make it public for applications who
really want to use it.

Note that chunked transfer encoding puts data through extra copies.
First it copies through the buffering area (because I have to be able
to extract variable length strings from inside the data stream), and then
again to reassemble the chunks into a single unified object.

We do assume that the user wants the entire thing as a single object.
This means that using this to pull unbounded data will just silently
consume all memory. Use caution!