move all public headers to include/nng/ folder
This change makes embedding nng + nggpp (or other projects depending on
nng) in cmake easier.  The header files are moved to a separate include
directory.  This also makes installation of the headers easier, and
allows clearer identification of private vs public heade files.

Some additional cleanups were performed by @gedamore, but the main
credit for this change belongs with @gregorburger.