fixes #823 Define public IPC
This introduces a basic IPC API, modeled on the TCP API, for direct access.
Only connection options are exposed at present -- we need to add options
for dialers and listeners (and particularly listener settings for
permissions and security attributes.)  Documentation is still outstanding,
but a very limited test suite exists.