fixes #1112 POSIX pollq finalizers could be simpler
We reap the connections when closing, to ensure that the clean up is
done outside the pollq thread.  This also reduces pressure on the
pollq, we think.  But more importantly it eliminates some complex
code that was meant to avoid deadlocks, but ultimately created other
use-after-free challenges.  This work is an enabler for further
simplifications in the aio/task logic.  While here we converted some
potentially racy locking of the dialers and reference counts to simpler
lock-free reference counting.