fixes #1154 perf programs could use options to select different protocols
This allows --reqrep0, --pair0, --pair1, and --bus0 to be used with
the inproc_lat, and also inproc_thr (though only with pair options for
that for now).  It also introduced --url for both programs to support
testing over different transports.

Also, we no longer pass out the header for REQ reply -- that is
an error, and led to some unfortunate failures when reusing the
message.