Windows: add a check for broken legacy environments without timespec_get.
If you want to build on Windows use a toolchain that supports modern APIs.
This means, for Microsoft, UCRT (Universal C Runtime), which is supported
by default on modern Visual Studio.  MinGW users may have to go out of their
way to enable it.  (New -D_UCRT flag or something.)

The supported toolchain for building on Windows is Visual Studio.  Use of
other tool chains is not officially supported or guaranteed to work.  YMMV.