posix: add getentropy() based RNG
XPG8 defines getentropy() as the only good source for random numbers.
However, real world use a bit more nuanced.  On BSD systems, we would
prefer to use arc4random as it avoids unnecessary system calls.  On
Linux however, getentropy is implemented in terms of getrandom, and should
be used directly when available.