Logging as a single chapter.
This is the direction we want to go with documentation.
Major areas of functionality should be one single chapter in
the API reference, which should make them far more usable
for people not already familiar with NNG.

This represents a substantial departure from UNIX man pages,
but many of our users are not used to working with UNIX man
pages anyway, and organization of man pages, while convenient
for reading with the legacy shell man command, fails on a number
of usability fronts.