fixes #1408 Reference count as a first class type
This starts by using this for the nni_pipe, but we will use it
for the other primary objects as well.  This should simplify
the tear down and hopefully eliminate some races.

It does mean that pipe destruction goes through an additional
context switch, for now at least.  This shouldn't be on the hot
data path anyway.