http server: sprintf is considered deprecated
Because it is typically associated with insecure code, use of sprintf
is discouraged.  Note that our usage was actually quite careful and
not insecure, but its mere presence raises concern especially by parties
who are unwilling or unable to assess the actual code for correctness.

A better choice here would be strlcat, but strlcat is not universally
available.