aio: significant rework
The attempt to use nni_task_abort() was completely misguided.
In fact this function isn't needed, and is a relic of a design that
predates the nni_aio_begin / nni_aio_schedule split.

Additionally, nni_aio_abort needed a fix to prevent a hang if
it was called between the calls to nni_aio_prep and nni_aio_schedule.
(Essentially a canceled operation should fail in scheduling.)