aio: run callback from aio_begin if stopped
There is a very subtle case where an aio can be closed, but we
don't realize it until precisely the moment of nni_aio_begin.
This is eaten, and the submitter winds up having no way of knowing
that the aio is closed, leading to a potential hang, or a leaked
resource.

It is very important that aio users check the return value and do
not resubmit an operation once NNG_ECLOSED is returned.  Otherwise
a circular loop can result.