windows ipc: significant refactor
This refactors a lot of the IPC code to hopefully address various
hangs on shutdown, etc.  The problem is that named pipes are not
terrifically reliable when it comes to aborting ConnectNamedPipe.
Additionally there were some logic errors in some of our code that
left things rather brittle.

Ultimately this all needs to be replaced with UNIX domain sockets
which are superior in many ways.