websocket transport: use inline SP objects
This also fixes a possible race in the listener that may cause
connections to be dropped incorrectly, if the connection arrives
before the common layer has posted an accept request.

Instead we save the connection and potentially match later, like
we do for the other protocols that need to negotiate.