posix pollers: expose pfd structures (for sizes) and fix poller selection
The poller selection in the previous poller changes for select were
not quite functional.  Also, while testing poll() based poller, there
were problems where it simply did not work correctly, so this addresses
those, and it seems to work now.

The pfd structures are exposed as we intend to allow inlining them
to eliminate the separate allocation and potential for failure during
initialization.  We also want to have plans afoot to eliminate a
lot of the extra locking done done on each I/O iteration, and this
is setting the foundation for that.