poll: performance improvements, simplifications
We preallocate the arrays used for pollfds, based on what the
system can tolerate (tunable with NNG_MAX_OPEN), and we change
the code for inserting and removing pollfds from the list so
that it can run without acquiring the locks during the main loop,
only when adding or removing files.

The poll() implementation is very nearly lock free in the hot
code path, and soon will be.