http: limit handler uri to 1K
This is just the part of the tree that will be matched when looking
up a handler.  Requests may come in with very much longer URIs, and
be matched to the handler as a "subdirectory".

This approach makes it possible to avoid a dynamic allocation on the
handler, at the cost of pre-allocating 1KB with the handler object.

This size can be overridden using a NNG_HTTP_MAX_URI at compile time.