tcp: flatten the listener implementation
The endpoints both use a nesting level for some common code and some
platform dependent code.  But the common code isn't that much and we
have similar patterns for e.g. IPC.

This avoids a layer of indirection in the structs, and extra allocations.
The payoff will be even larger for the dialers, but that is next.
(Dialers are more complicated because of DNS.)