TLS: break up the TLS layer a bit to refactor for DTLS.
This allows us to break the assumption that the bottom half is
TCP, or even an nng_stream, since the DTLS layer will use a totally
different layer.  Only nng_stream neeeds to support dial and listen.

Also: UDP: Make the sockaddr arguments to open const.

Also: Align the IPv6 address in the sockaddr (this allows for
efficient 64-bit or even 128-bit operations on these values.)